requirements.txt
torch
torchvision
matplotlib
scikit-learn
streamlit
matplotlib
pillow
tqdm
